use argparse + modularize upload
